requirements.txt
streamlit==1.36.0
plotly==5.10.0
pandas==2.2.2
numpy==1.23.1
folium==0.13.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
haversine==2.8.1
streamlit_folium==0.21.1
Pillow==9.2.0